requirements.txt
pyaudio>=0.2.11
opuslib>=3.0.1
numpy>=1.21.0
scipy>=1.7.0
websockets>=10.0
cryptography>=3.4.8
paho-mqtt>=1.6.1
boto3>=1.20.0
lgpio>=0.2.0

